rename nodejs watcher to be a library
